rename all files to central programs
