+++ src/components/central-programs-table.scss
