+++ src/pages/central-programs.js
@@ -2,7 +2,7 @@ import React from "react"
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import Table from "../components/departments-table"
+import Table from "../components/central-programs-table"
 import Sankey from "../components/sankey-chart"
 import { graphql } from "gatsby"
 import sankeyData from "../../data/sankey.json"
@@ -13,13 +13,13 @@ import "../styles/pages/central-programs.scss"
 
 const CentralProgramsPage = ({ data }) => {
 
-  const departments = data.allDepartmentsJson.nodes;
+  const centralPrograms = data.allCentralProgramsJson.nodes;
   return (
     <Layout pageClassName="central-programs-page">
-      <SEO title="Departments" />
+      <SEO title="Central Programs" />
       <h1>Central Programs Overview (2018-19)</h1>
       <Sankey data={sankeyData} restrictedData={sankeyRestrictedData} />
-      <Table data={departments} />
+      <Table data={centralPrograms} />
     </Layout>
   )
 }
@@ -27,10 +27,11 @@ const CentralProgramsPage = ({ data }) => {
 export default CentralProgramsPage
 
 export const query = graphql`
-  query DepartmentList {
-    allDepartmentsJson {
+  query CentralProgramsList {
+    allCentralProgramsJson {
       nodes {
         name
+        category
         budget
         spending
         year
